Render the float default as a type BigQuery has
The portable double is not portable: one dialect renders it as a type its
server does not have, and the failure arrives as a query error on that
backend alone. A precision-carrying float renders as double precision
where a precision is accepted, as the dialect's own double where it is
not, and as that server's 64-bit float on the dialect that rejected the
first attempt.

Oracle is the exception that has to stay one: naming a precision raises
there rather than compiling. It declares its own override for this type
already, so the shared default never reaches its dialect -- which makes
the two a pair, and the comment now says so, because removing the
override would surface the raise a long way from here.

The renderings are now written down per dialect. Nothing local can decide
whether a rendering is valid -- that is the server's opinion, and only a
run against one settles it -- but recording them means editing a default
changes a visible table in the same diff, so the per-backend consequences
are read while the change is being made rather than found afterwards.

+++ tests/integration/test_utils/data_source_config/sql.py
@@ -92,11 +92,16 @@ class SQLBatchTestSetup(BatchTestSetup[_SqlConfigT, TableAsset], ABC, Generic[_S
         default: InferrableTypesLookup = {
             str: sqltypes.VARCHAR,
             int: sqltypes.INTEGER,
-            # `Double`, not a bare DECIMAL: an unqualified decimal is read by several dialects
-            # as scale zero, which rounds every fractional value in a fixture frame to an
-            # integer on write, so an assertion about a value ends up asserting about a
-            # different one.
-            float: sqltypes.Double(),
+            # Not a bare DECIMAL: an unqualified decimal is read by several dialects as scale
+            # zero, which rounds every fractional value in a fixture frame to an integer on
+            # write, so an assertion about a value ends up asserting about a different one.
+            # Nor `Double`, which one dialect renders as a type its server does not have.
+            # A precision-carrying float resolves to double precision on the dialects that
+            # take a precision and to that dialect's own double elsewhere. It raises rather
+            # than compiling on Oracle, which is harmless only because Oracle declares its own
+            # override below -- removing that override would surface this, so the two belong
+            # together.
+            float: sqltypes.Float(precision=53),
             bool: sqltypes.BOOLEAN,
             date: sqltypes.DATE,
             # `DateTime`, not DATETIME or TIMESTAMP. Those two are emitted verbatim, and each is
